Fix tests after moving files.
