+++ src/engine/core/managers/treasures/TreasureManager.test.ts
@@ -5,8 +5,8 @@ import { registerActor, registry } from "@/engine/core/database";
 import { EGameEvent, EventsManager } from "@/engine/core/managers/events";
 import { TreasureManager } from "@/engine/core/managers/treasures/TreasureManager";
 import { ETreasureType, ITreasureDescriptor } from "@/engine/core/managers/treasures/treasures_types";
+import { giveInfo } from "@/engine/core/utils/info_portion";
 import { parseConditionsList } from "@/engine/core/utils/ini";
-import { giveInfo } from "@/engine/core/utils/object";
 import { ClientObject, ServerObject } from "@/engine/lib/types";
 import { mockActorClientGameObject, mockIniFile, mockServerAlifeObject } from "@/fixtures/xray";
