+++ src/engine/scripts/start.test.ts
@@ -7,7 +7,7 @@ import { registerExtensions } from "@/engine/scripts/register/extensions_registr
 import { registerManagers } from "@/engine/scripts/register/managers_registrator";
 import { registerSchemes } from "@/engine/scripts/register/schemes_registrator";
 
-jest.mock("@/engine/core/utils/class_ids", () => ({ updateClassIds: jest.fn(), createClassIds: jest.fn() }));
+jest.mock("@/engine/core/utils/class_ids_list", () => ({ updateClassIds: jest.fn(), createClassIds: jest.fn() }));
 jest.mock("@/engine/core/utils/ini/ini_system", () => ({ unlockSystemIniOverriding: jest.fn() }));
 jest.mock("@/engine/scripts/register/extensions_registrator", () => ({ registerExtensions: jest.fn() }));
 jest.mock("@/engine/scripts/register/managers_registrator", () => ({ registerManagers: jest.fn() }));
